requirements.txt
aiogram>=3.0.0
yt-dlp>=2023.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
asyncio>=3.4.3
psutil
aiofiles
time